requirements.txt
beautifulsoup4 
langchain 
sentence-transformers 
chromadb 
transformers 
torch
crawl4ai
streamlit
openai
accelerate
python-dotenv 
matplotlib 
nltk 
rouge-score
underthesea